display current weather data + fixes in data types

+++ Weather/Models/Current.swift
@@ -16,16 +16,16 @@ struct Current: Codable {
     var last_updated: String
     var temp_c: Float
     var temp_f: Float
-    var is_day: Int
+    var is_day: UInt8
     var condition: Condition
     var wind_mph: Float
     var wind_kph: Float
     var wind_degree: UInt16
     var wind_dir: String
-    var pressure_mb: UInt32
+    var pressure_mb: Float
     var pressure_in: Float
-    var precip_mm: UInt8
-    var precip_in: UInt8
+    var precip_mm: Float
+    var precip_in: Float
     var humidity: UInt8
     var cloud: UInt8
     var feelslike_c: Float
@@ -36,8 +36,8 @@ struct Current: Codable {
     var heatindex_f: Float
     var dewpoint_c: Float
     var dewpoint_f: Float
-    var vis_km: UInt8
-    var vis_miles: UInt8
+    var vis_km: Float
+    var vis_miles: Float
     var uv: Float
     var gust_mph: Float
     var gust_kph: Float
